Register tools in a new thread
Currently it can take a long time to load all the Galaxy tools into the toolbox. This change lets you load those tools in a new thread, so that it doesn't block kernel execution. This way you can continue to work on your notebook while you wait for them to load.